fix: move buildInfo into service
