+++ pkg/status/interfaces.go
@@ -7,4 +7,5 @@ import (
 type ServiceInterface interface {
 	KratosStatus(context.Context) bool
 	HydraStatus(context.Context) bool
+	BuildInfo(context.Context) *BuildInfo
 }
