+++ pkg/status/service.go
@@ -2,6 +2,7 @@ package status
 
 import (
 	"context"
+	"runtime/debug"
 
 	"github.com/canonical/identity-platform-login-ui/internal/healthcheck"
 	"github.com/canonical/identity-platform-login-ui/internal/logging"
@@ -11,6 +12,11 @@ import (
 	hClient "github.com/ory/hydra-client-go/v2"
 	kClient "github.com/ory/kratos-client-go"
 )
+
+type BuildInfo struct {
+	Version string `json:"version"`
+	Name    string `json:"name"`
+}
 
 type Service struct {
 	kratos kClient.MetadataApi
@@ -22,6 +28,23 @@ type Service struct {
 	tracer  trace.Tracer
 	monitor monitoring.MonitorInterface
 	logger  logging.LoggerInterface
+}
+
+func (s *Service) BuildInfo(ctx context.Context) *BuildInfo {
+	ctx, span := s.tracer.Start(ctx, "status.Service.BuildInfo")
+	defer span.End()
+
+	info, ok := debug.ReadBuildInfo()
+
+	if !ok {
+		return nil
+	}
+
+	buildInfo := new(BuildInfo)
+	buildInfo.Name = info.Main.Path
+	buildInfo.Version = s.gitRevision(ctx, info.Settings)
+
+	return buildInfo
 }
 
 func (s *Service) KratosStatus(ctx context.Context) bool {
@@ -76,6 +99,19 @@ func (s *Service) hydraReady(ctx context.Context) (bool, error) {
 	s.monitor.SetDependencyAvailability(tags, available)
 
 	return ok != nil, err
+}
+
+func (s *Service) gitRevision(ctx context.Context, settings []debug.BuildSetting) string {
+	ctx, span := s.tracer.Start(ctx, "status.Service.gitRevision")
+	defer span.End()
+
+	for _, setting := range settings {
+		if setting.Key == "vcs.revision" {
+			return setting.Value
+		}
+	}
+
+	return "n/a"
 }
 
 func NewService(kratos kClient.MetadataApi, hydra hClient.MetadataApi, tracer trace.Tracer, monitor monitoring.MonitorInterface, logger logging.LoggerInterface) *Service {
